Move WTRConfig class back out of test binary
